requirements.txt
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
plotly==5.19.0
psycopg2_binary==2.9.9
streamlit==1.32.0
